created stable package, created tests and configured pytest to run test from tests folder
