--- Pipfile
@@ -1,28 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-pytest = "*"
-awscli = "*"
-
-[packages]
-numpy = "*"
-scipy = "*"
-tensorflow = "==1.15.0"
-termcolor = "*"
-matplotlib = "*"
-pyyaml = "*"
-opencv-python = "*"
-pydot = "*"
-sklearn = "*"
-scikit-learn = "*"
-pipenv = "*"
-jsonschema = "*"
-jupyter = "*"
-notebook = "*"
-ipykernel = "*"
-
-[requires]
-python_version = "3.6"
