--- requirements.txt
@@ -1,18 +1,0 @@
-PyYAML
-jsonschema
-tensorflow
-matplotlib
-termcolor
-numpy
-scipy
-opencv-python
-# local package
--e .
-
-# external requirements
-click
-Sphinx
-coverage
-awscli
-flake8
-python-dotenv>=0.5.1
