+++ tox.ini
@@ -1,3 +1,3 @@
 [flake8]
-max-line-length = 79
+max-line-length = 120
 max-complexity = 10
